requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
lxml>=4.9.0
flask>=3.0.0
